refactor: move Button component to molecules folder
