+++ src/app/cart/page.tsx
@@ -1,6 +1,10 @@
-export default function CartPage() {
-  return (
+export default async function CartPage() {
+
+  return (  
     <div className="py-8">
+      <h1 className="font-bold text-xxl text-secondary md:text-xxxl">
+        Your Cart
+      </h1>
     </div>
   );
 }
