+++ src/components/molecules/productItem/index.tsx
@@ -1,5 +1,5 @@
 import Image from "next/image";
-import { Button } from "@/components/atoms/button";
+import { Button } from "@/components/molecules/button";
 import { Game } from "@/types/game";
 
 export interface ProductItemProps {
